pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tetris-rl"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "torch==2.4.0",
    "torchvision==0.19.0",
    "gymnasium==1.2.3",
    "tianshou==0.5.1",
    "numpy<2",
    "Pillow==12.1.1",
    "pygame",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
